Evaluation/chat-AstroLLaVA/pip-requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bitsandbytes==0.41.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.27.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
latex2mathml==3.77.0
llava-torch==1.2.2.post1
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.22.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.14
packaging==24.2
pandas==2.2.3
peft==0.4.0
pillow==10.4.0
psutil==6.1.1
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.2
safetensors==0.5.2
scikit-learn==1.2.2
scipy==1.15.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.15.0
tomlkit==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.36.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uv==0.5.18
uvicorn==0.34.0
wavedrom==2.0.3.post3
websockets==11.0.3
